requirements.txt
Flask
ak-minio
